requirements.txt
aioredis==1.3.0
alabaster==0.7.12
alembic==0.9.3.dev0
amqp==2.2.2
apturl==0.5.2
asgi-redis==1.4.3
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
Babel==2.7.0
billiard==3.5.0.3
blinker==1.4
Brlapi==0.6.6
cachetools==3.1.1
celery==4.1.0
certifi==2018.1.18
channels==2.0.0
channels-redis==2.2.1
chardet==3.0.4
Click==7.0
colorama==0.4.1
command-not-found==0.3
constantly==15.1.0
cryptography==2.1.4
cupshelpers==1.0
daphne==2.3.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==2.2.7
django-admin-sortable2==0.7.3
django-allauth==0.40.0
django-angular==2.2.3
django-appconf==1.0.3
django-cart==1.0.4
django-classy-tags==0.9.0
django-cms==3.7.0
django-cors-headers==2.1.0
django-countries==5.5
django-crispy-forms==1.8.0
django-entangled==0.3
django-filer==1.6.0
django-formtools==2.1
django-fsm==2.6.1
django-fsm-admin==1.2.4
django-haystack==2.5.1
django-imagekit==4.0.2
django-inline-actions==2.2.0
django-ipware==2.1.0
django-mathfilters==0.4.0
django-mptt==0.10.0
django-notifs==2.6.0
django-phonenumber-field==3.0.1
django-phonenumbers==1.0.1
django-polymorphic==2.0.3
django-post-office==3.2.1
django-rest-auth==0.9.5
django-sekizai==1.0.0
django-select2==6.3.1
django-shop==1.0.2
django-shop-ajax==0.1.3
django-starcross-gallery==1.0.6
django-state-field==0.3
django-templated-mail==1.0.0
django-treebeard==4.3
djangocms-admin-style==1.4.0
djangocms-cascade==1.1.5
djangocms-text-ckeditor==3.8.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djoser==1.1.5
docopt==0.6.2
docutils==0.15.2
easy-thumbnails==2.6
elasticsearch==5.0.1
feedparser==5.2.1
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-HTMLmin==1.3.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gevent==1.2.2
gitsome==0.8.4
google-api-python-client==1.4.1
google-auth==1.7.0
google-auth-httplib2==0.0.3
greenlet==0.4.12
hiredis==1.0.0
html5lib==1.0.1
htmlmin==0.1.12
httpie==1.0.3
httplib2==0.9.2
hyperlink==18.0.0
idna==2.7
imagesize==1.1.0
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
jsonfield==2.0.2
keyring==10.6.0
keyrings.alt==3.0
kombu==4.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
msgpack==0.5.6
msgpack-python==0.5.6
mysqlclient==1.4.5
netifaces==0.10.4
numpydoc==0.9.1
oauth==1.0.1
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.45.1
packaging==19.2
paramiko==2.0.0
passlib==1.7.1
pexpect==4.2.1
phonenumbers==8.10.20
pika==0.11.2
pilkit==2.0
Pillow==5.0.0
ply==3.11
prompt-toolkit==2.0.10
protobuf==3.0.0
psutil==5.4.2
psycopg2==2.8.4
psycopg2-binary==2.7.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.4.2
pygobject==3.26.1
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.6.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==2.4.5
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
python3-openid==3.1.0
pytz==2017.3
pyxdg==0.25
PyYAML==3.12
redis==2.10.6
reportlab==3.4.0
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.10
sqlparse==0.3.0
sshtunnel==0.1.4
stripe==2.40.0
system-service==0.3
systemd-python==234
Twisted==18.7.0
txaio==18.8.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.0.23
uritemplate==3.0.0
uritemplate.py==3.0.2
urllib3==1.22
usb-creator==0.3.3
uWSGI==2.0.15
vine==1.1.4
virtualenv==16.7.6
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
WTForms==2.1
xkit==0.0.0
zope.interface==4.5.0
